requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
pyautogui>=0.9.54
numpy>=1.24.0
pillow>=10.0.0
tkinter 